Remove _openedApplicationURL from CRWWebController
_openedApplicationURL is a set that is used by inserting the URL for
external apps that was launched on it. Then in HandleError if the error
was policy interruption for a URL inserted on the set ignore it.

The problem with that logic is HandleError is never called for requests
rejected by DecidePolicyForNavigationAction which is the place that
Chromium reject/allow app launches.

Bug: 850760
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Icc2535af2fef6ade9fdf28350c55336f8dbf51c1
Reviewed-on: https://chromium-review.googlesource.com/1124971
Reviewed-by: Danyao Wang <danyao@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#572312}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 20308970000824790df3c053c04fb7f5618be43b