[iOS] Support Reduce Motion in tag grid transitions.
This CL adds support for motion-reduced tab grid transitions.

The major change is to update the GridToHiddenTabAnimator to support
both presentation and dismissal, and rename it to ReducedMotionAnimator.

This animator is now used when reduced motion is enabled, as well as
for the prior case where the active tab isn't visible in the grid when
a tab is presented. The animation itself is unchanged.

Bug: 821047
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ie78a8c089f333e3767ee9f88c0c3283d9f82326e
Reviewed-on: https://chromium-review.googlesource.com/1127043
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#572954}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3c565bb768d327050ac2abe949bc243253ff6d35