[AF] Add autoselect suggestion bit when popup is shown on ARROW_DOWN
This is the first of a series of CLs to fix the interaction between
the autofill dropdown and Windows screenreader Jaws. When a dropdown
is shown using ARROW_DOWN, Jaws expects the first suggestion to be
announced; when that doesn't happen, it considers there are no items
to announce, so it captures ARROW_DOWN/ARROW_UP events to navigate
to the next/previous field on the page.

The fix consists of ensuring a suggestion is announced when the
dropdown is shown via ARROW_DOWN. There is no need to change behavior
when the dropdown is shown due on user clicks.

This CL propagates a bit from the rendered to the dropdown controller
indicating when the first suggestion must be selected on ARROW_DOWN
events. Please notice the bit is not set as true yet, but I'd like to
split this CL into smaller ones for easier review. This one touches
many files because I needed to propagate one bit from the renderer
to the autofill popup controller via several different paths.

The new behavior will be enabled on follow up CLs.

Bug: 848427
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ief5d5a47128b30fb1427231456f11659b865419c
Reviewed-on: https://chromium-review.googlesource.com/1124939
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Hiroshi Ichikawa <ichikawa@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#572972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 83662a20621fb84c00ee690b970bca3b574204a3