[ios] Disables History searchBar while editing.
This is a preparation CL to remove animations when populating
the history TableView and when performing a search.

In order achieve that we need to reload all the content of the
TableView instead of inserting/removing specific entries. This
causes previously selected items to deselect when searching.
For this reason there's no purpose to searching while editing
is enabled, furthermore selecting the search bar will deselect
selected items, which can be frustrating for the user.

This CL disables the searchBar while the table is on Edit mode.
The user can still search something and edit the results of the
query, but it can't perform a new search query before exiting
edit mode. This is the behavior of the legacy implementation.

Screenshots:
Enabled
https://drive.google.com/open?id=1ZdplK968U4XgBE1cFSB9L4scblJwFeBN
https://drive.google.com/open?id=1dH47e9tPHmDd2eUuOUVIt-W3Nay4v_A1

Disabled
https://drive.google.com/open?id=1Gp4IXNawohaZnlwSnQTSJ0LLbvaFW6Iw
https://drive.google.com/open?id=1aDPPjOd3DqX99vncP2wNXNHx_XKRyGhM

Bug: 859736
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I705a936f8bb339ab29a1431d7d4d969d77337d71
Reviewed-on: https://chromium-review.googlesource.com/1148890
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#577944}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6eadc4cb631d7726ba8ae001446bb945b0128c31