[iOS] Fix toolbar flash in grid transition.
This CL fixes the flash of grid visibility under the grid toolbars that happens during the grid transitions.

The fix has two parts. The first is to fade the grid toolbar backgrounds to black (masking out the other grid cells) instead of fading them to alpha zero. The toolbar controls still fade out, however.

The second part is to move the expanding tab in the view hierarchy so that it sits on top of the tab grid, while the rest of the animating cells are inserted inside it.

The API for this is somewhat clumsy at this stage; the animation class (itself a UIView) exposes the animating cell as a property, and the animator reparents it to be above the presenting view during the animation.

Bug: 865454
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I9ced094dfc053233f0dcbb5adb856639528c5daa
Reviewed-on: https://chromium-review.googlesource.com/1154974
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579805}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 481c3593b63e6cbf9e056714ae958cd8c42621cf