[iOS] Accessibility adjustments
This CL adds a delay for the voice over announcement of the InProduct
Help tips. This delay is added to prevent the announcement to be
cancelled by the element being focused, when the tab is presented.
It also adds an accessibility label to the empty space where the omnibox
should be if the toolbar was visible.

Bug: 866936
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ia75dd56772203e39b8d0bf1e94bef9cfacd87fdd
Reviewed-on: https://chromium-review.googlesource.com/1158236
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e910c7f7397ace03f42e1e39663de5ec93a9a58