[ios] Fix stale snapshotting and snapshots flickering blank
It is possible that the navigation completes in the time it takes to
animate to the tab grid. The navigation completion will trigger another
snapshot. But that snapshot may have the previous webpage because the
rendering is not guaranteed to have been completed at navigation
completion.

This CL implements a pausing/resuming mechanism so that we pause
snapshotting when animating to the tab grid while the webpage was
loading. We resume snapshotting once the animation to the tab grid
has completed.

This CL also fixes the snapshot going blank for a second while
animating to the tab grid.

Bug: 853569, 862799
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ib6f5cf1ca896c06940c7ce142518b94ffb3c7d37
Reviewed-on: https://chromium-review.googlesource.com/1156135
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579923}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 081eb1d1bad7865b3774941414ffa2a14eef7586