[iOS] Make LabelLink working with unbreakable spaces
When the LabelLink is displaying a text containing an unbreakable space,
the unbreakable space is preventing two words from being broken at the
end of a line. However, the ManualTextFramer used to get the position of
the link isn't respecting the unbreakable spaces.
So the displayed text isn't breaking the words at the end of the line
but the link is positioned as if the words were broken. So the position
of the link is wrong.

Bug: 872290
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I9b2bae32a47e19b14c397e230aade87b3cb88157
Reviewed-on: https://chromium-review.googlesource.com/1169014
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#582098}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b217ee7efcf09233819d5480d93c8c0085b9059