[ios] Pass in dependency into TabGridMediator
This CL fixes the failing TabGridMediator unittests.
The TabRestoreService is a dependency that is needed
in the TabGridMediator, but that requires a valid
browserState.
This CL properly passes in the TabRestoreService
dependency, instead of allowing the TabGridMediator
to grab it using the browserState.

Other dependencies should also be passed in, but
this CL is intentionally kept small.

Bug: 873632, 873633
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ie262f283b35c5a8cf5299d00e0a8cf680628cf53
Reviewed-on: https://chromium-review.googlesource.com/1176966
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#583607}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02d31f92255758524360faea590467f9bec56a58