[iOS] Fix positioning of the TabStrip LayoutGuide
The positioning of the LayoutGuide named kTabStripTabSwitcherGuide in
the TabStrip was wrong because the frame wasn't converted to the layout
guide owning view.
It leads to an incorrect positioning, as the width of the button and the
height of the status bar weren't taken into account.

Bug: 876642, 875760
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I537976c049508e3f0ac4a2e52ef654333a9ce1b3
Reviewed-on: https://chromium-review.googlesource.com/1184706
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#584981}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7fc059fcdda0de0f724fef8ce444835d295c3dad