[iOS] pre-cache tab grid images for appearance.
When there are more than six tabs in the tab grid, the animation when
entering the grid will sometimes be missing one or more snapshots. They
will appear a moment after the animation completes, which doesn't look
great.

The reason this happens is that the in-memory cache for snapshots is
capped at six objects. Requests for snapshots that are in the cache
return synchronously, but requests for uncached snapshots (which are
read from the filesystem) return in the next runloop. All of the layout
for the animation is done in the same runloop, so any uncached snapshots
aren't available.

The easiest solution is to increase the size of the snapshot cache, but
the number of snapshots that are needed is quite high in some cases (more
than 30 on a 12" iPad pro).

Instead, this CL implements a "prepare for appearance" mechanism for the
tab grid. This is triggered by a new ApplicationCommand that is sent when
the user touches *down* on a tab grid control. That leaves at least one
event loop for fetching cached resources.

The tab grid image data source (that is, the tab grid mediator) is ulti-
mately responsible for pre-loading the snapshots that might be needed. The
simple heuristic is to fetch the snapshots for each tab whose index is
within |N| of the active tab's index, where |N| is the maximum number of
tabs that will be visible in the grid. This is aggressive, but these pre-
fetched images are discarded as soon as the grid transition is completed,
and it's not as aggressive as just fetching all of the snapshots for the
open tabs.

Bug: 867980
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ie00e599af50cd01ba002139b017ba2c2ec864db7
Reviewed-on: https://chromium-review.googlesource.com/1190203
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#586682}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 55695ccf24a670fc62ccb2d62a636eec93c78573