[ios] Bookmarks TableVC uses dynamic row height.
This CL changes BookmarksTableVC so it uses dynamic row height.

It also updates some EG tests that started to fail once this change
was made. I suspect this has to do with internal TableView management
of cells when using dynamic height.

The BottomCell test cell was moved to the penultimate position since
we now snap at the top of the top cells instead
of using the offset value of the TableView. This meant that the top
cell was completely visible when re-creating the cache and the
"Bottom Cell" was partially hidden.



Bug: 869671
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: I3f02b580332161b3bd30b3df1cb7da0c782edabb
Reviewed-on: https://chromium-review.googlesource.com/1205806
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#589073}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fcd25fc0ff18ee08db307402282ba9748b0e4521