[Nav Experiment] Remove session restore from SetWebUsageEnabled.
WKBasedNavigationManagerImpl now has the ability to automatically
restore session if LoadIfNecessary() is called when the navigation
manager is in DETACHED mode (i.e. not having a WKWebView). As long as
all use cases of SetWebUsageEnabled(true) is followed by a
LoadIfNecessary(), the session restore in SetWebUsageEnabled() is
redundant. This is always the case in production code paths.

Restoring session in SetWebUsageEnabled() broke TabUsageRecorderTestCase
because this test toggles SetWebUsageEnabled() to simulate tab
eviction. Restoring session in SetWebUsageEnabled(true) triggers the
reload events of the evicted tab before SwitchToNormalMode() is called
and breaks metrics collection. This is not a problem in production
because WKWebView eviction is triggered by iOS system and does not
actually exercise the SetWebUsageEnabled() code path.

With this CL, all TabUsageRecorderTestCase pass again.

Bug: 881825
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: I81ebe71ff939d139f0eeca17ec0213bba1648447
Reviewed-on: https://chromium-review.googlesource.com/1225862
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Danyao Wang <danyao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#591817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc6e5024e81e73ca5a0c09d13a66027b7ed26b7c