[iOS] Check SideSwipe location in window coordinates
This CL changes the way the SideSwipe is checking if the touch location
is inside the toolbar frame or not by passing it in the window
coordinates instead of in the gesture's view coordinates.
This starts creating an issue as the toolbar frame is now contained in
a container, changing its frame's origin.

Bug: 883694
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: Iaf42b36ce01c47addd87ae71772abff605e75a1c
Reviewed-on: https://chromium-review.googlesource.com/1224387
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#592014}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92d8795b92e83d185aa52f4ad8b43c02d93e4639