[iOS][MF] Keyboard observer helper
Creates a helper that observes the keyboard. This observer is used to
more precisely remove the animations of the presented view, and also to
correctly restore the keyboard when it disappears. It is a separate
file to make it easier to test, in a separate CL.

This CL also cleans the logic for hiding subviews which is obsolete and
not needed.

This adds a new Earl Grey test suite, where in a follow up CL will
enable the manual fallback flags so they can be tested.

Bug: 845472
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: I299b36915d1d2b6f0ef642e73b302c9cd2f4f931
Reviewed-on: https://chromium-review.googlesource.com/1233835
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#594310}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97ec10ef30d024b391967066f22659be34325f1d