Fix a problem when storekit view is allocated and not presented.
Some websites try to open itunes app, then redirects immediatly to
http itunes url, in that case Chrome present app launch prompt and
while it's there itunes url is loaded, so it allocates a storekit view
and tries to present it and fails.
The problem is on subsequent itunes http navigations, the coordinator
check if there's a controller allocated, if so it will exit. And because
the one allocated in the previous case was never dismissed it breaks the
expected behavior.

Bug: 892503
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: Ie08160e37c6f27c02a081e06c9b956bc835cf4c7
Reviewed-on: https://chromium-review.googlesource.com/c/1263876
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#597355}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a375f958042e9e91e460f19587676369fe84006